# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dive Hub API",
    "version": "0.3.0",
    "description": "Sign in with POST /api/auth/sign-in/email (Better Auth); the session cookie authenticates every other call."
  },
...
  "paths": {
    "/api/setup": {
      "get": {
        "summary": "Whether this instance still needs its first admin",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "needed"
                  ],
                  "properties": {
                    "needed": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create the first admin with the setup token printed to the server log",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token",
                  "email",
                  "name",
                  "password"
                ],
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 16,
                    "maxLength": 128
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 254
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "password": {
                    "type": "string",
                    "minLength": 15,
                    "maxLength": 128,
                    "description": "At least 15 characters"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "email",
                    "name",
                    "role",
                    "disabled",
                    "createdAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "role": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "user"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "admin"
                          ]
                        }
                      ]
                    },
                    "disabled": {
                      "type": "boolean",
                      "description": "Disabled Users can't sign in; their data stays"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/invitations/lookup": {
      "post": {
        "summary": "The open invitation behind an invitation token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token"
                ],
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 16,
                    "maxLength": 128
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "email",
                    "expiresAt"
                  ],
                  "properties": {
                    "email": {
                      "type": "string"
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/invitations/accept": {
      "post": {
        "summary": "Accept an invitation: become a User and get signed in",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token",
                  "name",
                  "password"
                ],
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 16,
                    "maxLength": 128
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "password": {
                    "type": "string",
                    "minLength": 15,
                    "maxLength": 128,
                    "description": "At least 15 characters"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "email",
                    "name",
                    "role",
                    "disabled",
                    "createdAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "role": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "user"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "admin"
                          ]
                        }
                      ]
                    },
                    "disabled": {
                      "type": "boolean",
                      "description": "Disabled Users can't sign in; their data stays"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/me/preferences": {
      "patch": {
        "summary": "Change display settings (language, units); fields left out stay as they are",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "language": {
                    "anyOf": [
                      {
                        "type": "string",
                        "pattern": "^[a-z]{2,3}(-[A-Z]{2})?$",
                        "description": "BCP 47 tag, e.g. \"de\""
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "description": "Null: follow the browser"
                  },
                  "units": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "metric"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "imperial"
                        ]
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "description": "Null: follow the region of the browser"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "language",
                    "units"
                  ],
                  "properties": {
                    "language": {
                      "anyOf": [
                        {
                          "type": "string",
                          "pattern": "^[a-z]{2,3}(-[A-Z]{2})?$",
                          "description": "BCP 47 tag, e.g. \"de\""
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "Null: follow the browser"
                    },
                    "units": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "metric"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "imperial"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "Null: follow the region of the browser"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/invitations/{id}": {
      "delete": {
        "summary": "Revoke an open invitation (admins only)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "error"
                  ],
                  "properties": {
                    "code": {
                      "enum": [
                        "sign_in_required",
                        "admins_only",
                        "not_found",
                        "invalid_input",
                        "internal_error",
                        "upload_missing",
                        "upload_too_large",
                        "setup_done",
                        "setup_token_invalid",
                        "invitation_invalid",
                        "invitation_not_found",
                        "email_taken",
                        "reset_link_invalid",
                        "user_not_found",
                        "user_disabled",
                        "last_admin",
                        "not_yourself",
                        "confirmation_mismatch",
                        "session_not_found",
                        "import_not_found",
                        "dive_not_found",
                        "recording_not_found"
                      ],
                      "description": "Stable, machine-readable reason; clients translate it"
                    },
                    "error": {
                      "type": "string",
                      "description": "English description, for logs and scripts"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}